spotify-analysis/requirements.txt
spotipy
python-dotenv
pandas
numpy
matplotlib
